Network.MQTT.Auth

Documentation

class Credentials c where

This class defines how credentials that have been gathered from a connection attempt look like (at least). An Authenticator may use whatever information it finds suitable to authenticate the Principal.

Methods

credentialSecure :: c -> Bool

credentialUsername :: c -> Maybe Text

credentialPassword :: c -> Maybe ByteString

credentialCertificate :: c -> Maybe Certificate

credentialHttpHeaders :: c -> [(CI ByteString, ByteString)]

class Authenticator a where

Minimal complete definition

authenticate

Methods

authenticate :: Credentials c => a -> c -> IO (Maybe Principal)

data Principal

Constructors

Principal

Fields

principalIdentity :: ByteString

principalReceiveFilter :: RoutingTree ()

principalPublishFilter :: RoutingTree ()

principalSubscribeFilter :: RoutingTree ()