Network.MQTT.Authentication

Description

Documentation

class (Exception (AuthenticationException a), Eq (Principal a), Ord (Principal a), Show (Principal a)) => Authenticator a where

A peer identity optionally associated with connection/session specific information. newtype Principal = Principal T.Text deriving (Eq, Ord, Show)

An Authenticator is able to determine a Principal's identity from a Request.

Minimal complete definition

newAuthenticator, authenticate, hasPublishPermission, hasSubscribePermission

Associated Types

data Principal a

data AuthenticatorConfig a

data AuthenticationException a

This Exception may be thrown by any operation within this class. Operations must only throw this type of exception. Other exceptions won't be catched and may kill the broker.

Methods

newAuthenticator :: AuthenticatorConfig a -> IO a

Try to determine a Principal's identity from connection Request.

The operation shall return Nothing in case the authentication mechanism is working, but couldn't associate an identity. It shall throw and AuthenticationException in case of other problems.

authenticate :: a -> ConnectionRequest -> IO (Maybe (Principal a))

hasPublishPermission :: a -> Principal a -> Topic -> IO Bool

hasSubscribePermission :: a -> Principal a -> Filter -> IO Bool

data ConnectionRequest

This class defines how the information gathered from a connection request looks like. An Authenticator may use whatever information it finds suitable to authenticate the Principal.

Constructors

ConnectionRequest

Fields

requestClientIdentifier :: Text

requestCleanSession :: Bool

requestSecure :: Bool

Is this connection secure in terms of Transport Layer Security?

requestCredentials :: Maybe (Username, Maybe Password)

The username and password supplied with the MQTT handshake.

requestHttp :: Maybe (ByteString, [(CI ByteString, ByteString)])

The HTTP request head in case the client connected via WebSocket.

requestCertificateChain :: Maybe CertificateChain

An X.509 certificate chain supplied by the peer. It can be assumed that the transport layer implementation already verified that the peer owns the corresponding private key. The validation of the certificate claims (including certificate chain checking) must be performed by the Authenticator.

requestRemoteAddress :: Maybe ByteString