Network.MQTT.Authorization

Description

Documentation

class (Authenticator a, Exception (AuthorizationException a)) => Authorizer a where

An Authorizer is able to determine a Principal's permissions.

Associated Types

data AuthorizationException a

This Exception may be thrown by any operation within this class. Operations must only throw this type of exception. Other exceptions won't be catched and may kill the kill.

Methods

getPublishPermissions :: Principal a -> IO (RoutingTree ())

Determine the set of Topics a Principal is allowed to publish on.

Beware that this operation is called on each publication. The implementation is advised to cache the result (i.e. in an MVar) and update it from time to time.

getSubscribePermissions :: Principal a -> IO (RoutingTree ())

Determine the set of Filters a Principal is allowed to subscribe. The Principal is implicitly allowed to subscribe all filters that are more specific than those explicitly stated in the RoutingTree.

Beware that this operation is called on each subscription. The implementation is advised to cache the result (i.e. in an MVar) and update it from time to time.

Also note that Filters already subscribed by the client won't be unsubscribed if they are no longer in the current permission set. There is no way specified by MQTT to signal this to the client other than terminating the session.