Network.MQTT.Client

Description

Documentation

data Client t

data ClientConfiguration t

Constructors

ClientConfiguration

Fields

clientURI :: URI

mqtt://user:password@server.example:1234

clientWill :: Maybe Will

clientKeepAlive :: KeepAlive

clientIdentifierPrefix :: String

clientMaxUnacknowlegedMessages :: Word16

clientNewTransceiver :: IO t

Events

data ClientEvent

Constructors

Started

Connecting

Connected

Ping

Pong

Received Message

Disconnected ClientException

Stopped

listenEvents :: Client s -> IO (BroadcastListener ClientEvent)

acceptEvent :: BroadcastListener ClientEvent -> IO ClientEvent

Exceptions

data ClientException

Constructors

ProtocolViolation String

ConnectionRefused ConnectionRefusal

ClientLostSession

ServerLostSession

ClientClosedConnection

ServerClosedConnection

newClient :: ClientConfiguration t -> IO (Client t)

start :: (Connectable s, Address s ~ a, StreamConnection s, Closable s, Data s ~ ByteString) => Client s -> IO ()

Starts the Client if it is not already running.

The Started event will be added to the event stream.

The Connecting event will be added to the event stream.

A new Connectable will be created.

A connection will be tried to establish (includes connect as well as the MQTT handshake).

The Connected event will be added to the event stream.

Messages will be processed until a Disconnect occurs.

At every step after the Connecting event, a Disconnect may occur. The client will then try to reconnect automatically.

stop :: Closable s => Client s -> IO ()

Stops the Client if it is not already stopped.

A graceful DISCONNECT packet will be sent to the server and the connection will be closed.

All internal client threads will be terminated.

The Stopped event will be added to the event stream.

publish :: Client s -> Topic -> QoS -> Retain -> Payload -> IO ()

subscribe :: Client s -> [(TopicFilter, QoS)] -> IO [Maybe QoS]

unsubscribe :: Client s -> [TopicFilter] -> IO ()