Network.MQTT.RoutingTree

Description

RoutingTree

newtype RoutingTree a

The RoutingTree is a map-like data structure designed to hold elements that can efficiently be queried according to the matching rules specified by MQTT. The primary purpose is to manage client subscriptions, but it can just as well be used to manage permissions etc.

The tree consists of nodes that may or may not contain values. The edges are filter components. As some value types have the concept of a null (i.e. an empty set) the RoutingTreeValue is a class defining the data family RoutingTreeNode. This is a performance and size optimization to avoid unnecessary boxing and case distinction.

Constructors

RoutingTree (Map TopicFilterLevel (RoutingTreeNode a))

class RoutingTreeValue a where

Minimal complete definition

nodeNull, nodeTree, nodeValue, nodeFromTree, nodeFromTreeAndValue

Associated Types

data RoutingTreeNode a

Methods

nodeNull :: a -> Bool

nodeTree :: RoutingTreeNode a -> RoutingTree a

nodeValue :: RoutingTreeNode a -> Maybe a

nodeFromTree :: RoutingTree a -> RoutingTreeNode a

nodeFromTreeAndValue :: RoutingTree a -> a -> RoutingTreeNode a

null

null :: RoutingTree a -> Bool

empty

empty :: RoutingTree a

singleton

singleton :: RoutingTreeValue a => TopicFilter -> a -> RoutingTree a

matchTopic

matchTopic :: RoutingTreeValue a => Topic -> RoutingTree a -> Bool

Match a Topic against a RoutingTree.

The function returns true iff the tree contains at least one node that matches the topic and contains a value (including nodes that are indirectly matched by wildcard characters like + and # as described in the MQTT specification).

matchFilter

matchFilter :: RoutingTreeValue a => TopicFilter -> RoutingTree a -> Bool

Match a TopicFilter against a RoutingTree.

The function returns true iff the tree contains a path that is less or equally specific than the filter and the terminal node contains a value that is not nodeNull.

match (singleton "#") "a"     = True
match (singleton "#") "+"     = True
match (singleton "#") "a/+/b" = True
match (singleton "#") "a/+/#" = True
match (singleton "+") "a"     = True
match (singleton "+") "+"     = True
match (singleton "+") "+/a"   = False
match (singleton "+") "#"     = False
match (singleton "a") "a"     = True
match (singleton "a") "b"     = False
match (singleton "a") "+"     = False
match (singleton "a") "#"     = False

lookupWith

lookupWith :: RoutingTreeValue a => (a -> a -> a) -> Topic -> RoutingTree a -> Maybe a

insert

insert :: RoutingTreeValue a => TopicFilter -> a -> RoutingTree a -> RoutingTree a

insertWith

insertWith :: RoutingTreeValue a => (a -> a -> a) -> TopicFilter -> a -> RoutingTree a -> RoutingTree a

map

map :: (RoutingTreeValue a, RoutingTreeValue b) => (a -> b) -> RoutingTree a -> RoutingTree b

adjust

adjust :: RoutingTreeValue a => (a -> a) -> TopicFilter -> RoutingTree a -> RoutingTree a

delete

delete :: RoutingTreeValue a => TopicFilter -> RoutingTree a -> RoutingTree a

unionWith

unionWith :: RoutingTreeValue a => (a -> a -> a) -> RoutingTree a -> RoutingTree a -> RoutingTree a

differenceWith

differenceWith :: RoutingTreeValue a => (a -> a -> a) -> RoutingTree a -> RoutingTree a -> RoutingTree a