Network.Stack.Server

Description

Documentation

data TLS a

data WebSocket a

class Typeable a => ServerStack a where

Minimal complete definition

withServer, withConnection, flush, send, receive

Associated Types

type ServerMessage a

data Server a

data ServerConfig a

data ServerException a

data ServerConnection a

Methods

withServer :: ServerConfig a -> (Server a -> IO b) -> IO b

Creates a new server from a configuration and passes it to a handler function.

The server given to the handler function shall be bound and in listening state. The handler function is usually a forever loop that accepts and handles new connections.

withServer config $ \server->
  forever $ withConnection handleConnection

withConnection :: Server a -> (ServerConnection a -> IO b) -> IO (Async b)

Waits for and accepts a new connection from a listening server and passes it to a handler function.

This operation is blocking until the lowest layer in the stack accepts a new connection. The handlers of all other layers are executed within an Async which is returned. This allows the main thread waiting on the underlying socket to block just as long as necessary. Upper layer protocol handshakes (TLS etc) will be executed in a new thread.

withServer config $ \server-> forever $
  future <- withConnection server handleConnection
  putStrLn "The lowest layer accepted a new connection!"
  async $ do
      result <- wait future
      putStrLn "The connection handler returned:"
      print result

flush :: ServerConnection a -> IO ()

send :: ServerConnection a -> ServerMessage a -> IO ()

receive :: ServerConnection a -> Int -> IO (ServerMessage a)