Network.MQTT.Authentication

Documentation

class Authenticator a where

An instance of Authenticator is able to determine a Principals identity from authentication Credentials.

Minimal complete definition

authenticate

Methods

authenticate :: Credentials c => a -> c -> IO (Maybe Principal)

Try to determine a Principals identity from authentication Credentials.

class Credentials c where

This class defines how credentials that have been gathered from a connection request look like (at least). An Authenticator may use whatever information it finds suitable to authenticate the Principal.

Methods

credentialsSecure :: c -> Bool

Is this connection secure in terms of Transport Layer Security?

credentialsUsername :: c -> Maybe Text

The username supplied with the MQTT handshake.

credentialsPassword :: c -> Maybe ByteString

The password supplied with the MQTT handshake.

credentialsX509CertificateChain :: c -> [Certificate]

An X.509 certificate chain supplied by the peer. It can be assumed that the transport layer implementation already verified that the peer owns the corresponding private key. The validation of the certificate claims (including certificate chain checking) must be performed by the Authenticator.

credentialsWebSocketRequestHeaders :: c -> [(CI ByteString, ByteString)]

The HTTP request headers in case the client connected via WebSocket.

data Principal

The result of a call to authenticate.

Constructors

Principal

Fields

principalIdentity :: ByteString