Network.MQTT.Authentication

Description

Documentation

class (Show (Principal a), Exception (AuthenticationException a)) => Authenticator a where

An Authenticator is able to determine a Principal's identity from a Request.

Minimal complete definition

authenticate

Associated Types

data Principal a

A peer identity optionally associated with connection/session specific information.

data AuthenticationException a

This Exception may be thrown by any operation within this class. Operations must only throw this type of exception. Other exceptions won't be catched and may kill the broker.

Methods

authenticate :: Request r => a -> r -> IO (Maybe (Principal a))

Try to determine a Principal's identity from connection Request.

The operation shall return Nothing in case the authentication mechanism is working, but couldn't associate an identity. It shall throw and AuthenticationException in case of other problems.

class Request r where

This class defines how the information gathered from a connection request looks like. An Authenticator may use whatever information it finds suitable to authenticate the Principal.

Methods

requestSecure :: r -> Bool

Is this connection secure in terms of Transport Layer Security?

requestUsername :: r -> Maybe Text

The username supplied with the MQTT handshake.

requestPassword :: r -> Maybe ByteString

The password supplied with the MQTT handshake.

requestHeaders :: r -> [(CI ByteString, ByteString)]

The HTTP request headers in case the client connected via WebSocket.

requestCertificateChain :: r -> Maybe CertificateChain

An X.509 certificate chain supplied by the peer. It can be assumed that the transport layer implementation already verified that the peer owns the corresponding private key. The validation of the certificate claims (including certificate chain checking) must be performed by the Authenticator.