Network.Stack.Server

Description

Documentation

data TLS a

data WebSocket a

class Typeable a => ServerStack a where

Minimal complete definition

withServer, withConnection

Associated Types

data Server a

data ServerConfig a

data ServerException a

data ServerConnection a

data ServerConnectionInfo a

Methods

withServer :: ServerConfig a -> (Server a -> IO b) -> IO b

Creates a new server from a configuration and passes it to a handler function.

The server given to the handler function shall be bound and in listening state. The handler function is usually a forever loop that accepts and handles new connections.

withServer config $ \server->
  forever $ withConnection handleConnection

withConnection :: Server a -> (ServerConnection a -> ServerConnectionInfo a -> IO b) -> IO (Async b)

Waits for and accepts a new connection from a listening server and passes it to a handler function.

This operation is blocking until the lowest layer in the stack accepts a new connection. The handlers of all other layers are executed within an Async which is returned. This allows the main thread waiting on the underlying socket to block just as long as necessary. Upper layer protocol handshakes (TLS etc) will be executed in the new thread.

withServer config $ \server-> forever $
  future <- withConnection server handleConnection
  putStrLn "The lowest layer accepted a new connection!"
  async $ do
      result <- wait future
      putStrLn "The connection handler returned:"
      print result

class ServerStack a => StreamServerStack a where

Methods

sendStream :: ServerConnection a -> ByteString -> IO Int

sendStreamLazy :: ServerConnection a -> ByteString -> IO Int64

sendStreamBuilder :: ServerConnection a -> Int -> Builder -> IO Int64

receiveStream :: ServerConnection a -> Int -> IO ByteString

receiveStreamLazy :: ServerConnection a -> Int -> IO ByteString

class ServerStack a => MessageServerStack a where

This class is an abstraction for ServerStacks that support the transmission and reception of finite messages.

Minimal complete definition

sendMessage, sendMessages, receiveMessage, consumeMessages

Associated Types

type ClientMessage a

type ServerMessage a

Methods

sendMessage :: ServerConnection a -> ServerMessage a -> IO Int64

Send a message.

Returns the encoded message size.

sendMessages :: Foldable t => ServerConnection a -> t (ServerMessage a) -> IO Int64

Send several messages. This might lead to an improvement for very short messages.

Returns the summed size of all encoded messages.

receiveMessage :: ServerConnection a -> Int64 -> IO (ClientMessage a)

Receive a message.

The second parameter determines the maximum encoded message size which must not be exceeded by the client or an exception will be thrown. Implementations shall track the consumed bytes and shall throw an exception as soon as the limit is exceeded even if the message is not yet complete. This is important to prevent _denial of service_ attacks.

consumeMessages :: ServerConnection a -> Int64 -> (ClientMessage a -> IO Bool) -> IO ()

Consumes incoming messages with a supplied consumer callback.

The second parameter limits the size of a single encoded message (see receiveMessage).